Relocate misplaced callbacks
