+++ ui/src/shared/apis/query.ts
@@ -1,18 +1,12 @@
 import _ from 'lodash'
 import {getDeep} from 'src/utils/wrappers'
-import {
-  handleSuccess,
-  handleError,
-  handleLoading,
-} from 'src/shared/actions/timeSeries'
 import {analyzeQueries} from 'src/shared/apis'
 import {DEFAULT_DURATION_MS} from 'src/shared/constants'
 import replaceTemplates, {replaceInterval} from 'src/tempVars/utils/replace'
 import {proxy} from 'src/utils/queryUrlGenerator'
 
-import {Source} from 'src/types'
-
-import {Template} from 'src/types'
+import {Source, Status, Template} from 'src/types'
+import {TimeSeriesResponse, TimeSeriesSeries} from 'src/types/series'
 
 const noop = () => ({
   type: 'NOOP',
@@ -110,4 +104,60 @@ export const duration = async (
     console.error(error)
     throw error
   }
+}
+
+type EditQueryStatusFunction = (queryID: string, status: Status) => void
+
+const handleLoading = (
+  query: Query,
+  editQueryStatus: EditQueryStatusFunction
+): void =>
+  editQueryStatus(query.id, {
+    loading: true,
+  })
+
+const handleSuccess = (
+  data: TimeSeriesResponse,
+  query: Query,
+  editQueryStatus: EditQueryStatusFunction
+): TimeSeriesResponse => {
+  const {results} = data
+  const error = getDeep<string>(results, '0.error', null)
+  const series = getDeep<TimeSeriesSeries>(results, '0.series', null)
+  // 200 from server and no results = warn
+  if (!series && !error) {
+    editQueryStatus(query.id, {
+      warn: 'Your query is syntactically correct but returned no results',
+    })
+    return data
+  }
+
+  // 200 from chrono server but influx returns an "error" = warning
+  if (error) {
+    editQueryStatus(query.id, {
+      warn: error,
+    })
+    return data
+  }
+
+  // 200 from server and results contains data = success
+  editQueryStatus(query.id, {
+    success: 'Success!',
+  })
+  return data
+}
+
+const handleError = (
+  error,
+  query: Query,
+  editQueryStatus: EditQueryStatusFunction
+): void => {
+  const message =
+    getDeep<string>(error, 'data.message', '') ||
+    getDeep<string>(error, 'message', 'Could not retrieve data')
+
+  // 400 from chrono server = fail
+  editQueryStatus(query.id, {
+    error: message,
+  })
 }
